Improve in-memory database

+++ c2xcam/src/IDatabase.hpp
@@ -13,10 +13,10 @@ private:
 public:
   virtual ~IDatabase(){};
 
-  virtual bool exists(int id) = 0;
-  virtual DBView<CAM_t> get(int id) = 0;
-  virtual DBView<CAM_t> create(int id) = 0;
-  virtual void remove(int id) = 0;
+  virtual bool exists(long unsigned int id) = 0;
+  virtual DBView<CAM_t> get(long unsigned int id) = 0;
+  virtual DBView<CAM_t> create(long unsigned int id) = 0;
+  virtual void remove(long unsigned int id) = 0;
   virtual size_t entryCount() = 0;
 };
 }; // namespace c2x::cam
